Moved SqlConnectionGenerator
