+++ app/app/src/SqlDataStorageAndRetrival/SqlAccountService.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using app.Data;
 
 namespace app.src.SqlDataStorageAndRetrival
 {
