+++ app/app/src/SqlDataStorageAndRetrival/SqlPlaylistService.cs
@@ -4,6 +4,7 @@ using System.Data;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using app.Data;
 
 namespace app.src.SqlDataStorageAndRetrival
 {
